chore: alloy 0.4 (#11334)

+++ crates/primitives/src/transaction/compat.rs
@@ -20,7 +20,7 @@ impl FillTxEnv for TransactionSigned {
         tx_env.caller = sender;
         match self.as_ref() {
             Transaction::Legacy(tx) => {
-                tx_env.gas_limit = tx.gas_limit as u64;
+                tx_env.gas_limit = tx.gas_limit;
                 tx_env.gas_price = U256::from(tx.gas_price);
                 tx_env.gas_priority_fee = None;
                 tx_env.transact_to = tx.to;
@@ -34,7 +34,7 @@ impl FillTxEnv for TransactionSigned {
                 tx_env.authorization_list = None;
             }
             Transaction::Eip2930(tx) => {
-                tx_env.gas_limit = tx.gas_limit as u64;
+                tx_env.gas_limit = tx.gas_limit;
                 tx_env.gas_price = U256::from(tx.gas_price);
                 tx_env.gas_priority_fee = None;
                 tx_env.transact_to = tx.to;
@@ -48,7 +48,7 @@ impl FillTxEnv for TransactionSigned {
                 tx_env.authorization_list = None;
             }
             Transaction::Eip1559(tx) => {
-                tx_env.gas_limit = tx.gas_limit as u64;
+                tx_env.gas_limit = tx.gas_limit;
                 tx_env.gas_price = U256::from(tx.max_fee_per_gas);
                 tx_env.gas_priority_fee = Some(U256::from(tx.max_priority_fee_per_gas));
                 tx_env.transact_to = tx.to;
@@ -62,7 +62,7 @@ impl FillTxEnv for TransactionSigned {
                 tx_env.authorization_list = None;
             }
             Transaction::Eip4844(tx) => {
-                tx_env.gas_limit = tx.gas_limit as u64;
+                tx_env.gas_limit = tx.gas_limit;
                 tx_env.gas_price = U256::from(tx.max_fee_per_gas);
                 tx_env.gas_priority_fee = Some(U256::from(tx.max_priority_fee_per_gas));
                 tx_env.transact_to = TxKind::Call(tx.to);
@@ -76,7 +76,7 @@ impl FillTxEnv for TransactionSigned {
                 tx_env.authorization_list = None;
             }
             Transaction::Eip7702(tx) => {
-                tx_env.gas_limit = tx.gas_limit as u64;
+                tx_env.gas_limit = tx.gas_limit;
                 tx_env.gas_price = U256::from(tx.max_fee_per_gas);
                 tx_env.gas_priority_fee = Some(U256::from(tx.max_priority_fee_per_gas));
                 tx_env.transact_to = tx.to.into();
@@ -93,7 +93,7 @@ impl FillTxEnv for TransactionSigned {
             #[cfg(feature = "optimism")]
             Transaction::Deposit(tx) => {
                 tx_env.access_list.clear();
-                tx_env.gas_limit = tx.gas_limit as u64;
+                tx_env.gas_limit = tx.gas_limit;
                 tx_env.gas_price = U256::ZERO;
                 tx_env.gas_priority_fee = None;
                 tx_env.transact_to = tx.to;
