+++ crates/stages/stages/src/stages/execution.rs
@@ -293,7 +293,7 @@ where
                     target: "sync::stages::execution",
                     start = last_block,
                     end = block_number,
-                    throughput = format_gas_throughput((cumulative_gas - last_cumulative_gas) as u64, execution_duration - last_execution_duration),
+                    throughput = format_gas_throughput(cumulative_gas - last_cumulative_gas, execution_duration - last_execution_duration),
                     "Executed block range"
                 );
 
@@ -304,7 +304,7 @@ where
             }
 
             stage_progress = block_number;
-            stage_checkpoint.progress.processed += block.gas_used as u64;
+            stage_checkpoint.progress.processed += block.gas_used;
 
             // If we have ExExes we need to save the block in memory for later
             if self.exex_manager_handle.has_exexs() {
@@ -316,7 +316,7 @@ where
             if self.thresholds.is_end_of_batch(
                 block_number - start_block,
                 bundle_size_hint,
-                cumulative_gas as u64,
+                cumulative_gas,
                 batch_start.elapsed(),
             ) {
                 break
@@ -334,7 +334,7 @@ where
             target: "sync::stages::execution",
             start = start_block,
             end = stage_progress,
-            throughput = format_gas_throughput(cumulative_gas as u64, execution_duration),
+            throughput = format_gas_throughput(cumulative_gas, execution_duration),
             "Finished executing block range"
         );
 
@@ -459,7 +459,7 @@ where
                 stage_checkpoint.progress.processed -= provider
                     .block_by_number(block_number)?
                     .ok_or_else(|| ProviderError::HeaderNotFound(block_number.into()))?
-                    .gas_used as u64;
+                    .gas_used;
             }
         }
         let checkpoint = if let Some(stage_checkpoint) = stage_checkpoint {
@@ -572,7 +572,7 @@ fn calculate_gas_used_from_headers(
     let duration = start.elapsed();
     debug!(target: "sync::stages::execution", ?range, ?duration, "Finished calculating gas used from headers");
 
-    Ok(gas_total as u64)
+    Ok(gas_total)
 }
 
 /// Returns a `StaticFileProviderRWRefMut` static file producer after performing a consistency
@@ -760,7 +760,7 @@ mod tests {
                 total
             }
         }) if processed == previous_stage_checkpoint.progress.processed &&
-            total == previous_stage_checkpoint.progress.total + block.gas_used as u64);
+            total == previous_stage_checkpoint.progress.total + block.gas_used);
     }
 
     #[test]
@@ -801,7 +801,7 @@ mod tests {
                 total
             }
         }) if processed == previous_stage_checkpoint.progress.processed &&
-            total == previous_stage_checkpoint.progress.total + block.gas_used as u64);
+            total == previous_stage_checkpoint.progress.total + block.gas_used);
     }
 
     #[test]
@@ -834,7 +834,7 @@ mod tests {
                 processed: 0,
                 total
             }
-        }) if total == block.gas_used as u64);
+        }) if total == block.gas_used);
     }
 
     #[tokio::test]
@@ -925,7 +925,7 @@ mod tests {
                     }))
                 },
                 done: true
-            } if processed == total && total == block.gas_used as u64);
+            } if processed == total && total == block.gas_used);
 
             let provider = factory.provider().unwrap();
 
@@ -1077,7 +1077,7 @@ mod tests {
                         }
                     }))
                 }
-            } if total == block.gas_used as u64);
+            } if total == block.gas_used);
 
             // assert unwind stage
             assert_eq!(
