+++ examples/beacon-api-sidecar-fetcher/src/mined_sidecar.rs
@@ -116,7 +116,7 @@ where
                     let block_metadata = BlockMetadata {
                         block_hash: block.hash(),
                         block_number: block.number,
-                        gas_used: block.gas_used as u64,
+                        gas_used: block.gas_used,
                     };
                     actions_to_queue.push(BlobTransactionEvent::Mined(MinedBlob {
                         transaction,
@@ -195,7 +195,7 @@ where
                                         let block_metadata = BlockMetadata {
                                             block_hash: new.tip().block.hash(),
                                             block_number: new.tip().block.number,
-                                            gas_used: new.tip().block.gas_used as u64,
+                                            gas_used: new.tip().block.gas_used,
                                         };
                                         BlobTransactionEvent::Reorged(ReorgedBlob {
                                             transaction_hash,
@@ -268,7 +268,7 @@ async fn fetch_blobs_for_block(
                 let block_metadata = BlockMetadata {
                     block_hash: block.hash(),
                     block_number: block.number,
-                    gas_used: block.gas_used as u64,
+                    gas_used: block.gas_used,
                 };
                 BlobTransactionEvent::Mined(MinedBlob { transaction, block_metadata })
             })
